Fix value for SelectControl in block visibility  (#26300)
* Fix value for SelectControl in block visibility

Changes the variable used for the value attribute of the SelectControl used within the block visibility plugin to choose whether rules are Show or Hide conditions. The value was expecting an "action" attribute when the "action" variable is actually within the "conditions" attribute.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3101057581