Disconnect Dialog: Fix the prop types to prevent the React warnings (#26340)
The disconnect dialog was causing some warnings to be thrown up in the
browser dev console, when trying to disconnect Jetpack. This change
adjusts the prop types, and adds a key to the list in JetpackBenefits,
in order to prevent the warnings.

Co-authored-by: Karen Attfield <karenlattfield@gmail.com>
Co-authored-by: Samiff <samiff@users.noreply.github.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3114466297