Search: Add MockUsageMeter along with DonutMeterContainer component (#26344)
* Export DonutMeter from component library.

* First pass at unstyled output.

Generates the necessary HTML for the component.

* First pass at styling the component.

* Move component code into separate function.

Now it’s a component! We’re getting there…

* Add changelog.

* Added changelog for component library.

* Fixes for linting errors.

* New files for DonutMeterContainer.

Now inside “components/donut-meter-container” where it belongs.

* Introduce some sub-compenents.

* Build out InfoSecondary component.

Added props for the message string and the callback function to be called when the user clicks the link.

* Build out InfoPrimary component.

Added props for the message string and the callback function to be called when the user clicks the icon.

* Refactor handling of click events.

* .

* Add usage info to DonutMeterContainer

* Add i18n to DonutMeterContainer and title

* Introduce .upgrade-trigger section

* Add .record-meter-about section

* Refactor .jp-search-dashboard-meter-wrap styles

* Add the pricing gate to UsageMeter component

* Update changelog

* refactor numberWithCommas

Co-authored-by: oskosk <oskosk@gmail.com>
Co-authored-by: Dognose <dognose24@gmail.com>
Co-authored-by: Jasper Kang <kangzjnet@gmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3171490293