Jetpack Search: show free plan nudge to classic search users (#27372)
Before this change, classic search users
saw a CTA to Purchase Jetpack Instant Search,
but there was no way for them to start
using the free plan.

This commit changes that, and shows them
a button to try Jetpack Search for free
instead of the one for purchasing a plan.

It goes to checkout for jetpack_search_free
instead of jetpack_search now.

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3475540524